feat: pss-cln-auto-fold-service-001 - Relocate composed Automations root under internal
