+++ pkg/services/automations/internal/cron_watcher.go
@@ -1,4 +1,4 @@
-package service
+package internal
 
 import (
 	"context"
