+++ pkg/services/automations/internal/hosted_poller_test.go
@@ -1,4 +1,4 @@
-package service_test
+package internal_test
 
 import (
 	"context"
