+++ pkg/services/automations/internal/poller_watchers.go
@@ -1,4 +1,4 @@
-package service
+package internal
 
 import (
 	"context"
