+++ pkg/services/automations/internal/runtime_sidecars.go
@@ -1,4 +1,4 @@
-package service
+package internal
 
 import (
 	"context"
