+++ pkg/services/automations/internal/runtime_sidecars_test.go
@@ -1,4 +1,4 @@
-package service_test
+package internal_test
 
 import (
 	"context"
