+++ pkg/services/automations/internal/script_poller.go
@@ -1,4 +1,4 @@
-package service
+package internal
 
 import (
 	"context"
