+++ pkg/services/automations/internal/service_internal_test.go
@@ -1,4 +1,4 @@
-package service
+package internal
 
 import (
 	"context"
